fix: PR comments, factorize
